a lot of fixes
