+++ src/rest/handlers/presets.ts
@@ -4,9 +4,8 @@
 
 import { ItemsHandler } from '../base/items';
 import { ITransport } from '../transport';
-import { PresetType, DefaultType } from '../types';
-
-export type PresetItem<T = DefaultType> = PresetType & T;
+import { DefaultType } from '../../types';
+import { PresetItem } from '../../types';
 
 export class PresetsHandler<T = DefaultType> extends ItemsHandler<PresetItem<T>> {
 	constructor(transport: ITransport) {
