+++ src/rest/handlers/settings.ts
@@ -2,10 +2,8 @@
  * Settings handler
  */
 import { ITransport } from '../transport';
-import { SettingType, DefaultType } from '../types';
 import { SingletonHandler } from './singleton';
-
-export type SettingItem<T = DefaultType> = SettingType & T;
+import { SettingItem } from '../../types';
 
 export class SettingsHandler<T = SettingItem> extends SingletonHandler<T> {
 	constructor(transport: ITransport) {
