+++ src/rest/handlers/tfa.ts
@@ -1,5 +1,5 @@
 import { ITransport } from '../transport';
-import { TfaType, DefaultType } from '../types';
+import { TfaType, DefaultType } from '../../types';
 
 type TfaItem<T = DefaultType> = TfaType & T;
 export class TFAHandler {
