+++ src/rest/index.ts
@@ -3,11 +3,7 @@ export * from './auth';
 export * from './directus';
 export * from './handlers';
 export * from './items';
-export * from './singleton';
 export * from './transport';
 
 // Implementations
 export * from './base';
-
-// Types
-export * from './types';
