+++ src/websocket/base/directus.ts
@@ -1,76 +1,69 @@
 import { IDirectusWebSocket } from '../directus';
 import {
-	ActivityHandler,
-	AssetsHandler,
-	CollectionsHandler,
-	FieldsHandler,
-	FilesHandler,
-	FoldersHandler,
-	PermissionsHandler,
-	PresetsHandler,
-	RelationsHandler,
-	RevisionsHandler,
-	RolesHandler,
-	ServerHandler,
-	SettingsHandler,
-	UsersHandler,
-	UtilsHandler,
+	WebSocketActivityHandler,
+	WebSocketCollectionsHandler,
+	WebSocketFieldsHandler,
+	WebSocketFilesHandler,
+	WebSocketFoldersHandler,
+	WebSocketPermissionsHandler,
+	WebSocketPresetsHandler,
+	WebSocketRelationsHandler,
+	WebSocketRevisionsHandler,
+	WebSocketRolesHandler,
+	WebSocketSettingsHandler,
+	WebSocketUsersHandler,
 } from '../handlers';
-import { IItems } from '../items';
 import { IWebSocketTransport, WebSocketTransportOptions } from '../transport';
-import { ItemsHandler } from './items';
+import { WebSocketItemsHandler } from './items';
 import { WebSocketTransport } from './transport';
 import { WebSocketAuth } from './auth';
 import { LocalStorage, MemoryStorage } from '../../storage';
 import { TypeMap, TypeOf, PartialBy } from '../../types';
-import { GraphQLHandler } from '../handlers/graphql';
-import { ISingleton } from '../singleton';
-import { SingletonHandler } from '../handlers/singleton';
+import { ISingleton } from '../../items';
+import { WebSocketSingletonHandler } from '../handlers/singleton';
 import { DirectusStorageOptions } from '../../storage';
 import { IStorage } from '../../storage';
 import { IWebSocketAuth, WebSocketAuthOptions } from '../auth';
 import { Subscriptions } from './subscriptions';
+import { IItems } from '../items';
 
-
-export type DirectusOptions<IAuthHandler extends IWebSocketAuth = WebSocketAuth> = {
+export type WebSocketDirectusOptions<IAuthHandler extends IWebSocketAuth = WebSocketAuth> = {
 	auth?: IAuthHandler | PartialBy<WebSocketAuthOptions, 'transport' | 'storage'>;
 	transport?: IWebSocketTransport | Partial<WebSocketTransportOptions>;
 	storage?: IStorage | DirectusStorageOptions;
 };
 
-export class DirectusWebsocket<T extends TypeMap, IAuthHandler extends IWebSocketAuth = WebSocketAuth> implements IDirectusWebSocket<T> {
+export class DirectusWebsocket<T extends TypeMap, IAuthHandler extends IWebSocketAuth = WebSocketAuth>
+	implements IDirectusWebSocket<T>
+{
 	private _url: string;
-	private _options?: DirectusOptions;
+	private _options?: WebSocketDirectusOptions;
 	private _auth: IAuthHandler;
 	private _transport: IWebSocketTransport;
 	private _storage: IStorage;
-	private _assets?: AssetsHandler;
-	private _activity?: ActivityHandler<TypeOf<T, 'directus_activity'>>;
-	private _collections?: CollectionsHandler<TypeOf<T, 'directus_collections'>>;
-	private _fields?: FieldsHandler<TypeOf<T, 'directus_fields'>>;
-	private _files?: FilesHandler<TypeOf<T, 'directus_files'>>;
-	private _folders?: FoldersHandler<TypeOf<T, 'directus_folders'>>;
-	private _permissions?: PermissionsHandler<TypeOf<T, 'directus_permissions'>>;
-	private _presets?: PresetsHandler<TypeOf<T, 'directus_presets'>>;
-	private _relations?: RelationsHandler<TypeOf<T, 'directus_relations'>>;
-	private _revisions?: RevisionsHandler<TypeOf<T, 'directus_revisions'>>;
-	private _roles?: RolesHandler<TypeOf<T, 'directus_roles'>>;
-	private _users?: UsersHandler<TypeOf<T, 'directus_users'>>;
-	private _server?: ServerHandler;
-	private _utils?: UtilsHandler;
-	private _graphql?: GraphQLHandler;
-	private _settings?: SettingsHandler<TypeOf<T, 'directus_settings'>>;
+	private _activity?: WebSocketActivityHandler<TypeOf<T, 'directus_activity'>>;
+	private _collections?: WebSocketCollectionsHandler<TypeOf<T, 'directus_collections'>>;
+	private _fields?: WebSocketFieldsHandler<TypeOf<T, 'directus_fields'>>;
+	private _files?: WebSocketFilesHandler<TypeOf<T, 'directus_files'>>;
+	private _folders?: WebSocketFoldersHandler<TypeOf<T, 'directus_folders'>>;
+	private _permissions?: WebSocketPermissionsHandler<TypeOf<T, 'directus_permissions'>>;
+	private _presets?: WebSocketPresetsHandler<TypeOf<T, 'directus_presets'>>;
+	private _relations?: WebSocketRelationsHandler<TypeOf<T, 'directus_relations'>>;
+	private _revisions?: WebSocketRevisionsHandler<TypeOf<T, 'directus_revisions'>>;
+	private _roles?: WebSocketRolesHandler<TypeOf<T, 'directus_roles'>>;
+	private _users?: WebSocketUsersHandler<TypeOf<T, 'directus_users'>>;
+	private _settings?: WebSocketSettingsHandler<TypeOf<T, 'directus_settings'>>;
 	private _subscriptions: Subscriptions;
 
 	private _items: {
-		[collection: string]: ItemsHandler<any>;
+		[collection: string]: WebSocketItemsHandler<any>;
 	};
 
 	private _singletons: {
-		[collection: string]: SingletonHandler<any>;
+		[collection: string]: WebSocketSingletonHandler<any>;
 	};
 
-	constructor(url: string, options?: DirectusOptions) {
+	constructor(url: string, options?: WebSocketDirectusOptions) {
 		this._url = url;
 		this._options = options;
 		this._items = {};
@@ -98,7 +91,7 @@ export class DirectusWebsocket<T extends TypeMap, IAuthHandler extends IWebSocke
 			});
 		}
 
-		if(this._options?.auth instanceof IWebSocketAuth) {
+		if (this._options?.auth instanceof IWebSocketAuth) {
 			this._auth = this._options.auth;
 		} else {
 			this._auth = new WebSocketAuth({
@@ -107,8 +100,6 @@ export class DirectusWebsocket<T extends TypeMap, IAuthHandler extends IWebSocke
 				...this._options?.auth,
 			});
 		}
-
-		
 	}
 
 	get url() {
@@ -127,83 +118,84 @@ export class DirectusWebsocket<T extends TypeMap, IAuthHandler extends IWebSocke
 		return this._transport;
 	}
 
-	get assets(): AssetsHandler {
-		return this._assets || (this._assets = new AssetsHandler(this.transport));
+	get activity(): WebSocketActivityHandler<TypeOf<T, 'directus_activity'>> {
+		return (
+			this._activity || (this._activity = new WebSocketActivityHandler<TypeOf<T, 'directus_activity'>>(this.transport))
+		);
 	}
 
-	get activity(): ActivityHandler<TypeOf<T, 'directus_activity'>> {
-		return this._activity || (this._activity = new ActivityHandler<TypeOf<T, 'directus_activity'>>(this.transport));
-	}
-
-	get collections(): CollectionsHandler<TypeOf<T, 'directus_collections'>> {
+	get collections(): WebSocketCollectionsHandler<TypeOf<T, 'directus_collections'>> {
 		return (
 			this._collections ||
-			(this._collections = new CollectionsHandler<TypeOf<T, 'directus_collections'>>(this.transport))
+			(this._collections = new WebSocketCollectionsHandler<TypeOf<T, 'directus_collections'>>(this.transport))
 		);
 	}
 
-	get fields(): FieldsHandler<TypeOf<T, 'directus_fields'>> {
-		return this._fields || (this._fields = new FieldsHandler<TypeOf<T, 'directus_fields'>>(this.transport));
+	get fields(): WebSocketFieldsHandler<TypeOf<T, 'directus_fields'>> {
+		return this._fields || (this._fields = new WebSocketFieldsHandler<TypeOf<T, 'directus_fields'>>(this.transport));
 	}
 
-	get files(): FilesHandler<TypeOf<T, 'directus_files'>> {
-		return this._files || (this._files = new FilesHandler<TypeOf<T, 'directus_files'>>(this.transport));
+	get files(): WebSocketFilesHandler<TypeOf<T, 'directus_files'>> {
+		return this._files || (this._files = new WebSocketFilesHandler<TypeOf<T, 'directus_files'>>(this.transport));
 	}
 
-	get folders(): FoldersHandler<TypeOf<T, 'directus_folders'>> {
-		return this._folders || (this._folders = new FoldersHandler<TypeOf<T, 'directus_folders'>>(this.transport));
+	get folders(): WebSocketFoldersHandler<TypeOf<T, 'directus_folders'>> {
+		return (
+			this._folders || (this._folders = new WebSocketFoldersHandler<TypeOf<T, 'directus_folders'>>(this.transport))
+		);
 	}
 
-	get permissions(): PermissionsHandler<TypeOf<T, 'directus_permissions'>> {
+	get permissions(): WebSocketPermissionsHandler<TypeOf<T, 'directus_permissions'>> {
 		return (
 			this._permissions ||
-			(this._permissions = new PermissionsHandler<TypeOf<T, 'directus_permissions'>>(this.transport))
+			(this._permissions = new WebSocketPermissionsHandler<TypeOf<T, 'directus_permissions'>>(this.transport))
 		);
 	}
 
-	get presets(): PresetsHandler<TypeOf<T, 'directus_presets'>> {
-		return this._presets || (this._presets = new PresetsHandler<TypeOf<T, 'directus_presets'>>(this.transport));
+	get presets(): WebSocketPresetsHandler<TypeOf<T, 'directus_presets'>> {
+		return (
+			this._presets || (this._presets = new WebSocketPresetsHandler<TypeOf<T, 'directus_presets'>>(this.transport))
+		);
 	}
 
-	get relations(): RelationsHandler<TypeOf<T, 'directus_relations'>> {
-		return this._relations || (this._relations = new RelationsHandler<TypeOf<T, 'directus_relations'>>(this.transport));
+	get relations(): WebSocketRelationsHandler<TypeOf<T, 'directus_relations'>> {
+		return (
+			this._relations ||
+			(this._relations = new WebSocketRelationsHandler<TypeOf<T, 'directus_relations'>>(this.transport))
+		);
 	}
 
-	get revisions(): RevisionsHandler<TypeOf<T, 'directus_revisions'>> {
-		return this._revisions || (this._revisions = new RevisionsHandler<TypeOf<T, 'directus_revisions'>>(this.transport));
+	get revisions(): WebSocketRevisionsHandler<TypeOf<T, 'directus_revisions'>> {
+		return (
+			this._revisions ||
+			(this._revisions = new WebSocketRevisionsHandler<TypeOf<T, 'directus_revisions'>>(this.transport))
+		);
 	}
 
-	get roles(): RolesHandler<TypeOf<T, 'directus_roles'>> {
-		return this._roles || (this._roles = new RolesHandler<TypeOf<T, 'directus_roles'>>(this.transport));
+	get roles(): WebSocketRolesHandler<TypeOf<T, 'directus_roles'>> {
+		return this._roles || (this._roles = new WebSocketRolesHandler<TypeOf<T, 'directus_roles'>>(this.transport));
 	}
 
-	get users(): UsersHandler<TypeOf<T, 'directus_users'>> {
-		return this._users || (this._users = new UsersHandler<TypeOf<T, 'directus_users'>>(this.transport));
+	get users(): WebSocketUsersHandler<TypeOf<T, 'directus_users'>> {
+		return this._users || (this._users = new WebSocketUsersHandler<TypeOf<T, 'directus_users'>>(this.transport));
 	}
-	get settings(): SettingsHandler<TypeOf<T, 'directus_settings'>> {
-		return this._settings || (this._settings = new SettingsHandler<TypeOf<T, 'directus_settings'>>(this.transport));
-	}
-	get server(): ServerHandler {
-		return this._server || (this._server = new ServerHandler(this.transport));
-	}
-
-	get utils(): UtilsHandler {
-		return this._utils || (this._utils = new UtilsHandler(this.transport));
-	}
-
-	get graphql(): GraphQLHandler {
-		return this._graphql || (this._graphql = new GraphQLHandler(this.transport));
+	get settings(): WebSocketSettingsHandler<TypeOf<T, 'directus_settings'>> {
+		return (
+			this._settings || (this._settings = new WebSocketSettingsHandler<TypeOf<T, 'directus_settings'>>(this.transport))
+		);
 	}
 
 	singleton<C extends string, I = TypeOf<T, C>>(collection: C): ISingleton<I> {
 		return (
 			this._singletons[collection] ||
-			(this._singletons[collection] = new SingletonHandler<I>(collection, this.transport))
+			(this._singletons[collection] = new WebSocketSingletonHandler<I>(collection, this.transport))
 		);
 	}
 
 	items<C extends string, I = TypeOf<T, C>>(collection: C): IItems<I> {
-		return this._items[collection] || (this._items[collection] = new ItemsHandler<I>(collection, this.transport));
+		return (
+			this._items[collection] || (this._items[collection] = new WebSocketItemsHandler<I>(collection, this.transport))
+		);
 	}
 
 	subscribe<C extends string, I = TypeOf<T, C>>(collection: C, callback: (data: IItems<I>[]) => void) {
