+++ src/websocket/handlers/collections.ts
@@ -2,13 +2,10 @@
  * Collections handler
  */
 
-import { ManyItems, OneItem, ItemInput, QueryOne, EmptyParamError } from '../items';
+import { ManyItems, OneItem, ItemInput, QueryOne, EmptyParamError } from '../../items';
 import { IWebSocketTransport } from '../transport';
-import { CollectionType, DefaultType } from '../../types';
-
-export type CollectionItem<T = DefaultType> = CollectionType & T;
-
-export class CollectionsHandler<T = CollectionItem> {
+import { CollectionItem } from '../../types';
+export class WebSocketCollectionsHandler<T = CollectionItem> {
 	transport: IWebSocketTransport;
 	constructor(transport: IWebSocketTransport) {
 		this.transport = transport;
@@ -19,28 +16,30 @@ export class CollectionsHandler<T = CollectionItem> {
 		const response = await this.transport.request({
 			type: 'items',
 			collection: 'directus_collections',
-			action: 'get',
+			action: 'read',
 		});
 		return response.data as OneItem<NonNullable<T>>;
 	}
 
 	async readAll(): Promise<ManyItems<NonNullable<T>>> {
-		const response = await this.transport.request({
+		const response = await this.transport.request<ManyItems<NonNullable<T>>>({
 			type: 'items',
 			collection: 'directus_collections',
-			action: 'get',
+			action: 'read',
 		});
 
 		return response.data;
 	}
 
 	async createOne(collection: ItemInput<T>): Promise<OneItem<NonNullable<T>>> {
-		return (await this.transport.request<OneItem<NonNullable<T>>>({
-			type: 'items',
-			collection: 'directus_collections',
-			action: 'create',
-			data: collection,
-		})).data;
+		return (
+			await this.transport.request<OneItem<NonNullable<T>>>({
+				type: 'items',
+				collection: 'directus_collections',
+				action: 'create',
+				data: collection,
+			})
+		).data;
 	}
 
 	async createMany(collections: ItemInput<T>[]): Promise<ManyItems<NonNullable<T>>> {
@@ -51,7 +50,7 @@ export class CollectionsHandler<T = CollectionItem> {
 			data: collections,
 		});
 
-		return response.data
+		return response.data;
 	}
 
 	async updateOne(collection: string, item: ItemInput<T>, query?: QueryOne<T>): Promise<OneItem<NonNullable<T>>> {
