+++ src/websocket/handlers/fields.ts
@@ -2,13 +2,12 @@
  * Fields handler
  */
 
-import { ManyItems, OneItem, ItemInput, EmptyParamError, DefaultItem } from '../items';
+import { ManyItems, OneItem, ItemInput, EmptyParamError, DefaultItem } from '../../items';
 import { IWebSocketTransport } from '../transport';
-import { FieldType, DefaultType, ID } from '../../types';
+import { ID } from '../../types';
+import { FieldItem } from '../../types';
 
-export type FieldItem<T = DefaultType> = FieldType & T;
-
-export class FieldsHandler<T = FieldItem> {
+export class WebSocketFieldsHandler<T = FieldItem> {
 	transport: IWebSocketTransport;
 	constructor(transport: IWebSocketTransport) {
 		this.transport = transport;
@@ -41,7 +40,7 @@ export class FieldsHandler<T = FieldItem> {
 			},
 		});
 
-		return response.data as DefaultItem<NonNullable<T>>[];
+		return response.data;
 	}
 
 	async readAll(): Promise<ManyItems<NonNullable<T>>> {
@@ -50,35 +49,39 @@ export class FieldsHandler<T = FieldItem> {
 			collection: 'directus_fields',
 			action: 'read',
 		});
-		return response.data as DefaultItem<NonNullable<T>>[];
+		return response.data;
 	}
 
 	async createOne(collection: string, item: ItemInput<T>): Promise<OneItem<NonNullable<T>>> {
 		if (`${collection}` === '') throw new EmptyParamError('collection');
-		return (await this.transport.request<DefaultItem<NonNullable<T>>>({
-			type: 'items',
-			collection: 'directus_fields',
-			action: 'create',
-			data: {
-				...item,
-				collection,
-			},
-		})).data;
+		return (
+			await this.transport.request<DefaultItem<NonNullable<T>>>({
+				type: 'items',
+				collection: 'directus_fields',
+				action: 'create',
+				data: {
+					...item,
+					collection,
+				},
+			})
+		).data;
 	}
 
 	async updateOne(collection: string, field: string, item: ItemInput<T>): Promise<OneItem<NonNullable<T>>> {
 		if (`${collection}` === '') throw new EmptyParamError('collection');
 		if (`${field}` === '') throw new EmptyParamError('field');
-		return (await this.transport.request<DefaultItem<NonNullable<T>>>({
-			type: 'items',
-			collection: 'directus_fields',
-			action: 'update',
-			id: field,
-			data: {
-				...item,
-				collection,
-			},
-		})).data;
+		return (
+			await this.transport.request<DefaultItem<NonNullable<T>>>({
+				type: 'items',
+				collection: 'directus_fields',
+				action: 'update',
+				id: field,
+				data: {
+					...item,
+					collection,
+				},
+			})
+		).data;
 	}
 
 	async deleteOne(collection: string, field: string): Promise<void> {
