+++ src/websocket/handlers/files.ts
@@ -2,13 +2,12 @@
  * Files handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { FileType, DefaultType } from '../../types';
+import { DefaultType } from '../../types';
+import { FileItem } from '../../types';
 
-export type FileItem<T = DefaultType> = FileType & T;
-
-export class FilesHandler<T = DefaultType> extends ItemsHandler<FileItem<T>> {
+export class WebSocketFilesHandler<T = DefaultType> extends WebSocketItemsHandler<FileItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_files', transport);
 	}
