+++ src/websocket/handlers/folders.ts
@@ -2,13 +2,12 @@
  * Folders handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { FolderType, DefaultType } from '../../types';
+import { DefaultType } from '../../types';
+import { FolderItem } from '../../types';
 
-export type FolderItem<T = DefaultType> = FolderType & T;
-
-export class FoldersHandler<T = DefaultType> extends ItemsHandler<FolderItem<T>> {
+export class WebSocketFoldersHandler<T = DefaultType> extends WebSocketItemsHandler<FolderItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_folders', transport);
 	}
