+++ src/websocket/handlers/index.ts
@@ -1,5 +1,4 @@
 export * from './activity';
-export * from './assets';
 export * from './comments';
 export * from './collections';
 export * from './fields';
