+++ src/websocket/handlers/permissions.ts
@@ -2,13 +2,11 @@
  * Permissions handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { PermissionType, DefaultType } from '../../types';
+import { DefaultType, PermissionItem } from '../../types';
 
-export type PermissionItem<T = DefaultType> = PermissionType & T;
-
-export class PermissionsHandler<T = DefaultType> extends ItemsHandler<PermissionItem<T>> {
+export class WebSocketPermissionsHandler<T = DefaultType> extends WebSocketItemsHandler<PermissionItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_permissions', transport);
 	}
