+++ src/websocket/handlers/presets.ts
@@ -2,13 +2,12 @@
  * Presets handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { PresetType, DefaultType } from '../../types';
+import { DefaultType } from '../../types';
+import { PresetItem } from '../../types';
 
-export type PresetItem<T = DefaultType> = PresetType & T;
-
-export class PresetsHandler<T = DefaultType> extends ItemsHandler<PresetItem<T>> {
+export class WebSocketPresetsHandler<T = DefaultType> extends WebSocketItemsHandler<PresetItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_presets', transport);
 	}
