+++ src/websocket/handlers/relations.ts
@@ -1,12 +1,12 @@
 /**
  * Relations handler
  */
-import { ManyItems, OneItem, ItemInput, EmptyParamError } from '../items';
+import { ManyItems, OneItem, ItemInput, EmptyParamError } from '../../items';
 import { IWebSocketTransport } from '../transport';
-import { RelationType, DefaultType, ID } from '../../types';
+import { ID } from '../../types';
+import { RelationItem } from '../../types';
 
-export type RelationItem<T = DefaultType> = RelationType & T;
-export class RelationsHandler<T = RelationItem> {
+export class WebSocketRelationsHandler<T = RelationItem> {
 	transport: IWebSocketTransport;
 
 	constructor(transport: IWebSocketTransport) {
@@ -22,12 +22,12 @@ export class RelationsHandler<T = RelationItem> {
 			action: 'read',
 			id,
 		});
-		return response.data as OneItem<T>;
+		return response.data;
 	}
 
 	async readMany(collection: string): Promise<ManyItems<T>> {
 		if (`${collection}` === '') throw new EmptyParamError('collection');
-		const response = await this.transport.get({
+		const response = await this.transport.request({
 			type: 'items',
 			collection: 'directus_relations',
 			action: 'read',
@@ -52,33 +52,37 @@ export class RelationsHandler<T = RelationItem> {
 	}
 
 	async createOne(item: ItemInput<T>): Promise<OneItem<T>> {
-		return (await this.transport.request<OneItem<T>>({
-			type: 'items',
-			collection: 'directus_relations',
-			action: 'create',
-			data: item,
-		})).data;
+		return (
+			await this.transport.request<OneItem<T>>({
+				type: 'items',
+				collection: 'directus_relations',
+				action: 'create',
+				data: item,
+			})
+		).data;
 	}
 
 	async updateOne(collection: string, field: string, item: ItemInput<T>): Promise<OneItem<T>> {
 		if (`${collection}` === '') throw new EmptyParamError('collection');
 		if (`${field}` === '') throw new EmptyParamError('field');
-		return (await this.transport.request<OneItem<T>>({
-			type: 'items',
-			collection: 'directus_relations',
-			action: 'update',
-			data: item,
-			query: {
-				filter: {
-					collection: {
-						_eq: collection,
-					},
-					field: {
-						_eq: field,
+		return (
+			await this.transport.request<OneItem<T>>({
+				type: 'items',
+				collection: 'directus_relations',
+				action: 'update',
+				data: item,
+				query: {
+					filter: {
+						collection: {
+							_eq: collection,
+						},
+						field: {
+							_eq: field,
+						},
 					},
 				},
-			}
-		})).data;
+			})
+		).data;
 	}
 
 	async deleteOne(collection: string, field: string): Promise<void> {
@@ -97,7 +101,7 @@ export class RelationsHandler<T = RelationItem> {
 						_eq: field,
 					},
 				},
-			}
+			},
 		});
 	}
 }
