+++ src/websocket/handlers/revisions.ts
@@ -2,13 +2,12 @@
  * Revisions handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { RevisionType, DefaultType } from '../../types';
+import { DefaultType } from '../../types';
+import { RevisionItem } from '../../types';
 
-export type RevisionItem<T = DefaultType> = RevisionType & T;
-
-export class RevisionsHandler<T = DefaultType> extends ItemsHandler<RevisionItem<T>> {
+export class WebSocketRevisionsHandler<T = DefaultType> extends WebSocketItemsHandler<RevisionItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_revisions', transport);
 	}
