+++ src/websocket/handlers/roles.ts
@@ -2,13 +2,12 @@
  * Roles handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { WebSocketItemsHandler } from '../base/items';
 import { IWebSocketTransport } from '../transport';
-import { RoleType, DefaultType } from '../../types';
+import { DefaultType } from '../../types';
+import { RoleItem } from '../../types';
 
-export type RoleItem<T = DefaultType> = RoleType & T;
-
-export class RolesHandler<T = DefaultType> extends ItemsHandler<RoleItem<T>> {
+export class WebSocketRolesHandler<T = DefaultType> extends WebSocketItemsHandler<RoleItem<T>> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_roles', transport);
 	}
