+++ src/websocket/handlers/settings.ts
@@ -1,13 +1,11 @@
 /**
  * Settings handler
  */
+import { SettingItem } from '../../types';
 import { IWebSocketTransport } from '../transport';
-import { SettingType, DefaultType } from '../../types';
-import { SingletonHandler } from './singleton';
+import { WebSocketSingletonHandler } from './singleton';
 
-export type SettingItem<T = DefaultType> = SettingType & T;
-
-export class SettingsHandler<T = SettingItem> extends SingletonHandler<T> {
+export class WebSocketSettingsHandler<T = SettingItem> extends WebSocketSingletonHandler<T> {
 	constructor(transport: IWebSocketTransport) {
 		super('directus_settings', transport);
 	}
