+++ src/websocket/handlers/singleton.ts
@@ -1,23 +1,21 @@
 import { IWebSocketTransport } from '../transport';
-import { QueryOne, OneItem, ItemInput } from '../items';
-import { ISingleton } from '../singleton';
+import { QueryOne, OneItem, ItemInput } from '../../items';
+import { ISingleton } from '../../items';
 
-export class SingletonHandler<T> implements ISingleton<T> {
+export class WebSocketSingletonHandler<T> implements ISingleton<T> {
 	protected collection: string;
 	protected transport: IWebSocketTransport;
-	protected endpoint: string;
 
 	constructor(collection: string, transport: IWebSocketTransport) {
 		this.collection = collection;
 		this.transport = transport;
-		this.endpoint = collection.startsWith('directus_') ? `/${collection.substring(9)}` : `/items/${collection}`;
 	}
 
 	async read<Q extends QueryOne<T>>(query?: Q): Promise<OneItem<T, Q>> {
 		const item = await this.transport.request<OneItem<T, Q>>({
 			type: 'items',
 			collection: this.collection,
-			action: 'get',
+			action: 'read',
 			query,
 		});
 		return item.data;
