+++ src/websocket/index.ts
@@ -2,12 +2,8 @@
 export * from './auth';
 export * from './directus';
 export * from './handlers';
-export * from './items';
-export * from './singleton';
+export * from '../items';
 export * from './transport';
 
 // Implementations
 export * from './base';
-
-// Types
-export * from '../types';
